Larger cleanup - Move setting default "wordlist" from open_wordlists into bogoconfig.c   and bogoutil.c - Take word_lists argument for wordlists_base.? functions as first steps   of getting rid of the "word_lists" global. This also simplifies some   functions. - Drop some dead initializations. - Remove some dead code. - Drop calc_prob in favor of the calc_prob_pure function that does not   read global variables. - Move spam/good message count to local layers in several places. The   global msgs_good/msgs_bad are still needed in some places. - bogoutil now reads the .MSG_COUNT token in the same transaction as the   counts, for accuracy in concurrent transactions. - Drop set_msg_counts in many places. - Manually inline wordprob_* in score.c to make it more concise and   remove traces of the old split wordlist format. - Move some functions up in the module to get rid of a prototype.   There is seldom need for advance declarations of static functions.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@5010 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/wordhash.c
@@ -241,7 +241,7 @@ static void display_node(hashnode_t *n, const char *str)
 {
     wordprop_t *p = (wordprop_t *)n->buf;
     if (verbose > 2)
-	printf( "%20.20s %5d %5d%s", n->key->text, (int)p->cnts.bad, (int)p->cnts.good, str);
+	printf( "%20.20s %5u %5u%s", n->key->text, p->cnts.bad, p->cnts.good, str);
 }
 
 /* this function accumulates the word frequencies from the src hash to
